Improve extension download error in install report
When the extension installer can't download from any of the mirrors,
make the error that's written to the installation report more clear.

Previously the cloudfront mirror download would report `{:error,
{:error, :closed}}` as the installation failure reason, but this isn't
very descriptive.

I've wrapped the error in a more descriptive message so we can see where
the error actually comes from and what's happening. This format is
similar to how our Ruby gem formats this type of error:
https://github.com/appsignal/appsignal-ruby/blob/409fbf98e3974317243ae85a6f2e6e77132db88b/ext/base.rb#L142-L151

As far as I know the cloudfront mirror will always fail currently, so
this only happens if the fastly mirror (the first one it tries), also
fails.

Part of #683