Remove scheduling from Appsignal.Probes
This commit modifies the register/2 and unregister/1 functions in
Appsignal.Probes so that, when a function-based probe is
registered, it is automatically wrapped in a FunctionProbeWrapper
server-based probe and registered with the new dynamic
supervisor.

Before this change, Appsignal kept a map between probe names
and their corresponding probe functions. This is now a map
between probe names and their corresponding PIDs as they are
registered in the dynamic supervisor, preserving the user-facing
interface provided by the register/2 and unregister/1 functions.

Since functions are now wrapped into servers, started as
dynamic supervisor children and executed by the new scheduler, the
existing scheduling logic in Appsignal.Probes is now unused and
can now be removed.