Simplify Appsignal.Probes and add tests
This commit removes the logic that ensures the input of register/2
is a function, since it can now take the shape of any sort of
supervisor child spec.

It also includes some refactoring of Appsignal.Probes and its
tests, and it adds some tests regarding the behaviour of the
system when a second probe is registered under the same name.
The behaviour should be the same as it was before the changes,
that is, the new probe replaces the old one, but now it is
under test coverage.