Add scheduler utilization gauge to probe
This commit adds gauges corresponding to the scheduler utilization
metrics to Appsignal.Probes.ErlangProbe.

To obtain the scheduler utilization metrics corresponding to the
time range between the points in time A and B, one must call
:scheduler.sample/0 at time A, keep hold of that value, then
call :scheduler.utilization/1 at time B with the sample obtained
from time A.

Therefore, ErlangProbe needs to keep the previous sample in its
state, and to be able to do this, it is converted from a
function-based probe to a server-based one.

An integration test is added to ensure that the ErlangProbe, and
its new state-dependent metric, work correctly with the new
scheduler.

In `mix.exs`, `:runtime_tools` is added as a dependency. This
gets rid of a warning when using functions from `:scheduler`.