Fix SSL options download in installer for mirrors
When installer can't download the install package from the fastly mirror
it falls back on the cloudfront mirror since PR #671. On newer OTP
versions the cloudfront mirror fails to download the error message
`{:error, :closed}` from hackney, which isn't very descriptive.

After some searching around and trying out different things, I found
that removing the `ciphers` option made the cloudfront mirror download
work. After some more research with the help of @luismiramirez we found
this blog post: https://nts.strzibny.name/hackney-ssl-configuration/
that lists the options I decided to implement in this commit. Added to
the list of versions is also `tlsv1.3` as was previously configured in
the ciphers options as well.

This `versions` option seems like a simpler way to configure these
ciphers on newer OTP versions, than calling `:ssl.cipher_suites` with
different arguments for different OTP versions.

This method doesn't work on older OTP versions, which is why the
original implementation was kept in tact for OTP 22 and lower.

By passing in the `ssl_options` to hackney's `request` function we
override the OTP defaults for SSL, which we then need to set again for
the specific OTP version.

I've updated the mix helper and transmitter to both use the same config,
and use the same config for all OTP versions. I've testing this on the
CI with both the fastly and cloudfront mirror as the download and they
both pass.

This commit itself doesn't test the cloudfront mirror, because it's not
the first mirror used and the first mirror doesn't fail to download. In
the PR for this commit I'll link a CI build that does only use the
cloudfront mirror for testing.

Part of #683