Turn FakeFunctionProbe into an Agent
There is little benefit to `FakeFunctionProbe` being a
`TestAgent`, and the specificity of the `TestAgent`'s `child_spec`
gets in the way of spawning multiple `FakeFunctionProbe`s.

This commit makes `FakeFunctionProbe` into a regular `Agent`, and
uses it in the probe override test, which requires two different
probes to be register. This test previously used a vanilla `Agent`
as a workaround for not being able to spawn multiple
`FakeFunctionProbe` instances.