Address "function `ciphers/0` is unused" warning
As part of fixing the SSL options when downloading from mirrors
in the installer (#688), the TLS-related options passed to the
HTTP client were simplified for OTP 23 and above, no longer
requiring the use of the `ciphers/0` function. However, this
function was still defined under OTP 23, leading to a warning
about the function being unused.

This commit addresses the warning by only defining `ciphers/0`
for OTP versions below 23.