Fix diagnose files last 10 lines rendering
It looked like the `appsignal.log` section only rendered 9 lines, but it
was actually 10. The `appsignal.log` file included an empty line at the
end of the file.

I've trimmed that line so that it doesn't add unexpected empty lines in
the diagnose report and create unnecessary space between it and the next
section.