Support CLI diagnose options for releases
When an Elixir app is made into a release with `mix release` it's more
difficult to run the diagnose command. A user needs to eval this
command:

```
bin/your_app eval ":appsignal_tasks.diagnose()"
```

It's not possible to pass along `--send-report` this way, because it
will be will ignored for some reason.

To allow for the options to be available for this method of calling the
diagnose command, allow the Mix task (that's being called this way) to
receive arguments and pass them to the Diagnose module, like how it
would be done if one would call `mix appsignal.diagnose --send-report`.

This change is necessary for the diagnose tests
(https://github.com/appsignal/diagnose_tests) test suite, because we
need to run the diagnose from a released app. If we'd run a unreleased
app the AppSignal package would recompile on every run, making it
impossible to overwrite the install and download reports for testing.