Implement scheduler utilization metric
This commit implements the Erlang scheduler utilization metric
on the minutely probes.

In order to report on the scheduler utilization, the probe must
keep state from its previous execution. That is, to measure the
scheduler utilization between points in time A and B, you must
call `:scheduler.sample/0` on point A, then pass the sample it
returns to `:scheduler.utilization/1 on point B.

To be able to keep state in the probe, we keep a map of states
in `Appsignal.Probes` alongside the map of probes. If a probe
takes an argument, the state returned by its previous run is
retrieved from the states map and passed as the argument, and
the probe's return value is stored in the states map to be used
as the next run's state.