Revert "Add scheduler utilization gauge to ErlangProbe"
When reviewing the changes in this PR, I noticed that one of the
tests had accidentally been broken as part of a refactor, meaning
that failing probes were effectively not being tested. After
fixing this test and reintroducing failing probes in the mix,
the issues that I was seeing before were popping up again: both
regarding unregistering a probe while it is being restarted,
and regarding the failing probes making their supervisor reach
maximum restart intensity and bringing the supervisor down
with them.

I think it's likely that these issues, in practice, only affect
the tests, due to probes running in the tests every ten
milliseconds, instead of every sixty seconds like in non-test
environments. Regardless, I don't want to pollute the tests'
output or introduce flaky tests, and I wouldn't want to leave the
behavior for failing probes untested either.