Update `credo` to version 1.5.6
This commit updates the version of the `credo` linter to `1.5.6`.
It also addresses the new warnings given by the linter, most of
which relate to the use of `Application.get_env/3` in module
attributes, which are evaluated at compile time. Using
`Application.compile_env/3` instead removes the warning.

Since `compile_env` is not available in Elixir versions below
1.10, a helper macro is added to `Appsignal.Utils` and used to
conditionally call `get_env` or `compile_env` depending on the
Elixir version.

In order to use this macro from within the `mix_helpers.exs` file,
its code is duplicated within that file as `Mix.Appsignal.Utils`.