Do not print private diagnose_endpoint option
In the diagnose report we do not need to print the diagnose_endpoint
config option. It's a private option that users do not need to
customize. It's only there for testing purposes.

We can hide it in the server side viewer as well, but removing from the
sent report is a bit trickier, because the `Diagnose.Config.config/0`
function's return value is also used to configure where the report is
sent.