Include extra fields in install report
The install report in other integrations include the dependencies and
flags fields. This is a little more difficult to track for the Elixir
integration because they're set in the `Makefile`.

But let's standardize the install report format some more so we can more
easily add values here in the future.