Standardize key validation request
Add the other query params to the auth request.

Update the error message to be standard for all integrations. Not all
integrations necessarily have the response status code on error, so
make sure all integrations start with "Failed to validate".

Change the validation request to a POST request, like the other
integrations. Both work, but it's good to all use the same request
method.

Add test for diagnose Push API key output and report.
Update the diagnose tests with tests for the Push API key validation in
the diagnose CLI.