Deprecate and replace set_sample_data
This commit deprecates the `Appsignal.Span.set_sample_data/3`
method, replacing it with the `set_tags/1`, `set_params/1`,
`set_environment/1`, `set_session_data/1` and `set_custom_data/1`
methods.

Before sending the data to the extension, allow-list filtering is
applied if the `:phoenix, :filter_parameters` config key is set
to `{:keep, keys}`, where `keys` is a list of keys to allow
through. Before this change, this filtering would be applied to
any data passing through `set_sample_data/3`. After this change,
this filtering is not applied in any of the new helpers; the
Phoenix integration will be changed in a separate pull request to
apply this filtering to the parameters before passing them to
`set_params/1`.