Use only supported TLS versions
Ask the SSL module for a list of supported TLS versions for Hackney
to attempt to use during the agent download, so that versions that
are unsupported by Elixir's currently linked "cryptolib" (that is,
OpenSSL) are not listed in the `:versions` option.

This fixes #714, an issue where downloading the agent will fail
when using OpenSSL 1.0.2, as the download attempts to use the
TLS versions listed in `:versions` _in order_, and errors out when
it attempts to use TLS 1.3, which isn't supported by the cryptolib.

The same changes to the TLS options used in the agent download are
also applied to `Appsignal.Transmitter`, which is used to validate
API keys and send the diagnose report.