Use `ERLANG_VERSION` in CI tests
When running tests on Semaphore, different combinations of
Erlang/OTP versions and Elixir versions are tested. This is
implemented by calling `sem-version` with the desired Erlang or
Elixir version.

However, when `sem-version elixir` is called with an Elixir
version that is not installed, it will install and activate the
highest Erlang version supported by that Elixir version, as part
of the installation process. Since `sem-version erlang` is called
before `sem-version elixir`, the desired Erlang version is
effectively overriden.

This commit reverses the order in which `sem-version erlang` and
`sem-version elixir` are called.