Fix ignored Dialyzer warnings
For some reason, running Dialyzer on OTP 24 makes it correctly
identify the location in the source code where the warnings it is
set to ignore take place. This commit changes the entries in
`dialyzer.ignore-warnings` to match the warnings emitted with the
correct location, making the CI pass succeed in ignoring them.