Sample schedulers when probe starts
The scheduler utilization sample was being obtained only after all
the other metrics had been collected, and no second sample was
being passed to `:scheduler.utilization`, meaning the elapsed time
between samples was inconsistent. This commit modifies the Erlang
probe so that the sample is collected when the probe starts, and
so that the utilization is compared between the previous sample
and the next.