Only filter params sample data
Before this change, all sample data was passed through the
`Appsignal.Utils.MapFilter.filter/1` function, which filters the
given map by the allowlist specified in the `:filter_parameters`
Phoenix configuration key, when it exists and is an allowlist.

This commit changes the behaviour of the `set_sample_data/3`
function so that it only filters the sample data when the given
key is `"params"`. This fixes #649.