Unset the default log_level in the integration (#740)
If the log_level was set by default by the Elixir integration it would
always use that level, even if `debug` or `transaction_debug_mode` were
set. These options are deprecated, but if set, and `log_level` is not
configured by the app, we should listen to these options.

I've removed the default log_level from the integration, but it will be
set by the extension and agent's log_level behavior. Since there's no
centralized logger in the Elixir integration yet, we do not need
to update any configuration for the integration's logger.