Change how version is retrieved for diagnose
In the diagnose report, the agent version was obtained from the version
file that is generated after the agent installation is done.

To comply with the other integrations and to know which agent version is
being installed in the diagnose if something goes wrong, the version is
now obtained from `agent.exs` in the diagnose report.