Add send_session_data config option
This replaces the `skip_session_data` option that is now deprecated.

As discussed in https://github.com/appsignal/integration-guide/pull/114
the `skip_session_data` is inconsistently named compared to our other
options that do the same thing, but start with `send_`.

The behavior of `send_session_data` is also the exact opposite of
`skip_session_data`. It's on by default and if turned off it won't send
the session data.

Only when `send_session_data` is not set by the user, will
`skip_session_data` be listened to. If `send_session_data`'s value is
determined by the `skip_session_data_backwards_compatibility` method,
its value is set in the `system` source, because the system figured
out what value it needs to be set to. If the user has manually
configured it, it will not set it in the system source.

To keep appsignal-plug old versions compatibility, the
`skip_session_data` option is still set.

In the future, when the `skip_session_data` option is removed, the
`send_session_data` option should be added to the `@default_config`
attribute in the `Appsignal.Config` module and this behaviour should all
be removed.