Use log_level instead of debug
To match the changes implemented in [appsignal-agent#730][agent],
the implementation of `Appsignal.Config.debug?/0`, which is in
turn used by `Appsignal.Logger.debug/1`, is modified to use valid
`log_level` config values as its source of truth, and fall back on
the deprecated `debug` and `transaction_debug_mode` config options
only if no valid `log_level` value is present.

Before this change, this method only checked the `debug`
config option, not `transaction_debug_mode`. In order to be
consistent with the changes to the agent (`transaction_debug_mode`
implies `debug`, because a log level of "trace" implies "debug")
this method now checks `transaction_debug_mode` as well.

Fixes #746.

[agent]: https://github.com/appsignal/appsignal-agent/pull/730/