Add "override" config source
When a config option is set by the integration based on other config
options, the new config option values are written to the "override"
source.

This may happen for renamed config options that set the new config
option when the deprecated config option is set, like the
`skip_session_data` and `send_session_data` config options this commit
updates to this new source.

This "override" config source is leading. It should not be confused with
"system" that determines _defaults_ for the config based on environment
details like host and system (not AppSignal environment variables).

I've changed the order how the config sources and options are loaded in
the code, sources are only written to the `Application` env at the very
end, but it shouldn't change the result. This just made it easier to
modify the sources during the process.

Closes https://github.com/appsignal/appsignal-elixir/issues/752