Support arm64 host for tests
When I run the test suite on an arm64 host, support that architecture as
well in the test suite. I had to update on test that always assumed an
x86_64 architecture for the test suite.

I added a FakeErlang module to always return the x86_64 architecture.
It's only used in one test and seemed like the way to do it, otherwise
I'd have to construct the correct system architecture string again.

[skip changeset]