Implement compliant centralised logger
Rewrite the existing `Appsignal.Logger` to comply with the specs
in our integration guide, specifically regarding the log levels,
the log message format, logging to a file, and logging to stderr.

Fixes #747, although this commit does not actually make it so the
logger is used. Most logging statements throughout the application
still use Elixir's `Logger`.

See [the integration guide specification][ig] for more information.

[ig]: https://github.com/appsignal/integration-guide/blob/main/build/logging.md