Improve error reporting on extension load failure
It's a common problem for apps first installed on the host system
(macOS) and then mounted with the `deps` and `_build` directory in a
Docker container that's running another Operating System like Ubuntu to
not start the extension. This is because the extension is not installed
for the container's architecture.

If the target and architecture mismatch, print the installed and current
target and architecture so users can spot the difference between them.
This will hopefully make the error more clear: the user needs to
reinstall the extension for their host's architecture.

## Message printed on start

This error message will be printed on the start of the AppSignal
application and not on compilation. On compilation it will still print
the generic message:

> Error loading NIF (Is your operating system (#{arch}) supported?

I initially tried to add the message here, but parsing the JSON
`install.report` file caused errors with the `Appsignal.JSON` module
being undefined.

```
13:35:51.744 [error] Process #PID<0.1551.0> raised an exception
** (UndefinedFunctionError) function Appsignal.Json.decode/1 is undefined (module Appsignal.Json is not available)
    Appsignal.Json.decode("Contents of install.report file")
    lib/appsignal/nif.ex:42: Appsignal.Nif.fetch_installed_architecture_target/0
    lib/appsignal/nif.ex:23: Appsignal.Nif.init/0
    (kernel 8.2) code_server.erl:1317: anonymous fn/1 in :code_server.handle_on_load/5
```

I considered writing the architecture and target to a file named
`priv/architecture_target` as plain text and reading that there instead.
In the end I decided to move this check to the AppSignal initialization
whenever AppSignal is started. This is usually when people would care
for this message, so it's not logged in environments where AppSignal is
not active.

## Using the AppSignal logger

I changed the call from `Logger.error` to `Appsignal.logger.error` so
it's logged to both the log file and printed to STDERR with the agreed
upon format.

## Testing changes

To make FakeOS work, I had to configure it in `config/config.exs`,
otherwise it would be loaded too late and not apply to the `Appsignal`
module. I used the key `os_internal`, otherwise it would conflict with
the `mix_helpers.exs`'s `os` key for the OS and break the entire
installation in the test env.

I also had to add a fallback to the original `:os` module, because the
Appsignal application is also started when the test process starts,
causing it to break on start of `mix test`.