Fix release_upgrade test after config_change
The config_change function no longer returns a PID. Because of that we
don't know the PID of the new spawned process to monitor.

Alternatively the config_change function can return a tuple of
`{:ok, pid}`, but I'm not sure if that would work. I'm unfamiliar with
the return value this function, but the docs it only returns `:ok`.

https://hexdocs.pm/elixir/1.13.4/Application.html#c:config_change/3