Fix release_upgrade_test to match new behavior
The release_upgrade_test was broken in because it depended on the PID
for the spawned process being returned. In
b049d588166f52e5e9284e1098095560072885b9, this behavior was changed to
_always_ return :ok, as defined in the Application specification.

The test ensured the configuration was properly reloaded, but also
tested the side effect of the separate process being spawned, which is
an implementation detail.

Instead, we can test that the configuration is updated after a
reasonable amount of time (5 seconds, in this case). This tells us the
process was spawned properly. We can assume the process is also closed,
as it immediatly exits when the initialize/0 function is called.