Always return the span from Span setter functions
The set_name/2, set_namespace/2 and set_sample_data/3 functions all had
a rescue clause that returned a nil value, even if a span was passed to
them.

This caused issues with optional arguments, where a value that should be
a string was a nil. Instead of returning the span unchanged (which is
expected when chainging functions), these functions would then return a
nil, breaking the setters for the rest of the chain.

This patch makes sure the span is returned as passed, if it's a nil or a
proper span.