Update OTP & Erlang versions on CI (#789)
* Generate semaphore.yml from semaphore.yml.erb

    erb -T- .semaphore/semaphore.yml.erb > .semaphore/semaphore.yml

* Add Elixir 1.14.0 to CI

Elixir 1.14 is compatible with OTP 23--25. This patch adds it to OTP 23
and 24, 25 is added later.

* Put newest OTP version in variable in CI matrix

Instead of having to update the newest Erlang version throughout the
file, this patch adds an NEWEST_OTP variable that automatically takes
the first key from the VERSIONS list.

* Always run linters on newest Elixir version

Although the version was previously updated regularly, this patch makes
sure the linters always run in the newest released version of Elixir.

* Run No NIF test on released Elixir version

Instead of running the no-nif test on main, run it on the latest
released version of Elixir, which should be slightly faster.

* Add Elixir main to versions list

Instead of hardcoding main into the CI matrix, add it to the versions
list so we can run the tests on main over multiple OTP versions later.

* Update credo to ~> 1.6

The previous version didn't run on Elixir 1.14

* Use Enum.map_join/3 in appsignal.diagnose.ex

Credo remarked that using map_join/3 was is more efficient than using
map/2, then join/2.

* Remove unused span_id and trace_id delegates

Since the trace_id and span_id functions are currently never used, and
not exposed throught the Nif, this patch removes them from Test.Nif as
well.

* Add OTP 25 to CI

OTP 25 is compatibile with Elixir 1.14 and 1.13.

* Move VERSIONS to versions.rb

To allow using the defined versions from the other libraries, move the
versions list to a separate file.

* Add CI matrix generation scripts

The script/generate_ci_matrix script runs the erb command that generates
the matrix. The validate_ci_matrix makes sure the output matches the
file on disk. The latter is used in a check on CI.

* Add DO NOT EDIT note to build matrix

> # DO NOT EDIT
> # This is a generated file by the `script/generate_ci_matrix` task.
> # See `.semaphore/semaphore.yml.erb` for the build matrix.
> # Generate this file with `script/generate_ci_matrix`.

* Add note about updating the build matrix to README

Outline updating the file and running the script.