Fix undefined symbols specified for linking (#793)
We made a mistake in PR #792. We split up the very long line in the
Makefile to specify a symbol per line. This accidentally broke how the
linker interprets these options.

Specify the symbols per line, but on each line also include the `-Wl,`
prefix to specify the liker option. Somehow the multi line option value
for `-Wl,` didn't work.