Fix Alpine Linux ARM64 extension linking
The extension linking failed on symbols being defined multiple times.
This is because the Rust compiler includes the `__fe_getround`
`__fe_raise_inexact` [symbols for ARM64 targets][symbols].

I couldn't find a way to mark either symbol as "weak", so it would be
overridden with the other "strong" symbol. I couldn't find a way to mark
them as weak outside of their original definition.

Instead this change allows symbols to be defined multiple times. As per
the GNU libc [documentation of this linker flag][ld man], it will only
use the first defined symbol. (I link to the GNU linker docs, because I
can't find the musl linker docs.)

> --allow-multiple-definition
> -z muldefs
>   Normally when a symbol is defined multiple times, the linker will
>   report a fatal error. These options allow multiple definitions and
>   the first definition will be used.

Since I don't have a reliable way to detect Alpine Linux/musl systems in
the Makefile, this new flag applies to all ARM64 hosts. It won't
negatively affect builds in which this is not a problem (no multiple
definitions of a symbol).

Fixes #795

[symbols]: https://github.com/rust-lang/compiler-builtins/blob/83b2c07a6bfa394fa5979ca12b60536f402a792c/build.rs#L436-L437
[ld man]: https://www.man7.org/linux/man-pages/man1/ld.1.html