Fix FunctionClauseError for Finch <0.12 (#798)
Finch 0.11 and below report `telemetry` events with the same name
as those by Finch 0.12 and above, but the events carry different
meanings and the metadata map provided has a different structure.

Due to the different structure of the metadata map, our event
handlers resulted in a `FunctionClauseError`, causing the handlers
to detach.

This commit changes the event handlers so that events not matching
the expected structure are silently ignored. This means that we
do not break applications using Finch 0.11 and below, but those
Finch versions remain unsupported by our Finch integration.