Add error_backend_enabled configuration option (#806)
In order to allow users to disable the error backend, which has now been
replaced by our integrations, this patch adds a configuration option. By
default, the backend is enabled, to not change the current
functionality.

Users who get double or unneeded errors reported in the background
namespace can disable the error backend. The error backend will be
disabled by default in a future version before it's removed altogether
in the next major release.