Add Oban instrumentation
Instrument the `[:oban, :engine, :insert_job, ...]` telemetry span
events and create an AppSignal child span with data about the job
inserted. Use the the worker name as the span name, if available.

Instrument the `[:oban, :job, ...]` telemetry span-like events and
create an AppSignal root span with data about the job being
processed. Use the arguments given to the job as params, and set
useful attributes of the job, such as its ID, the queue, the attempt
number, its priority and the execution result, as tags. Use the job
worker as the name of the root span.

Set an error for the `[:oban, :job, :exception]` event.

Track a distribution gauge for the job duration, by worker, and
for each worker by state.

Track a distribution gauge for the job queue time, measured as the time
between scheduling the job and attempting it for currently starting
jobs, by queue.

Track a counter for when a job is inserted, by worker.

Track a counter for when a job is started, by worker and by queue.

Track a counter for when a job is stopped, by worker, and for each
worker by state. Exceptions are tracked as stops with a failure
state.