Support all v2.x versions
Instead of using the event structure implemented in later versions
of Oban, use the event structure from v2.0.0, which is kept in later
versions for backwards compatibility, as the base expectation for
the shape of the event.

When possible, use the information added in later versions to enrich
the span, report additional metrics or improve the reporting of
existing metrics.

Fix the handling of the `:tags` metadata property (which was being
treated as a map, not as a list of strings) and handle the `:meta`
job property similarly to how `:tags` was handled.