Don't update state in overridden probe test
For the test that checks if a probe's state is overridden, we check if
the state matches the second one passed. We do that by setting the state
to 10, and then setting it to 0.

Previously, the function passed was the same as in the other tests,
meaning it incremented the state for every tick. Since the probes are
running during the tests, there was a small chance the state was updated
after setting the second state and before the assertion.

This patch passes a function that returns an atom, but does not update
the state.