Don't match econnrefused in ReportTest exactly (#819)
One of the error scenarios in the diagnose reports is when an unexpected
error occurs. In the tests, this is simulated by taking the test
diagnose server down, resulting in a :econnrefused.

In some situations, the error thrown is a :closed, which made this test
brittle. To fix that, the test no longer explicitly tests for a
:econnrefused. Rather, it just tests if an error with a reason key is
returned when this happens.