Update MonitorTest to fix brittleness (#816)
* Remove nonexisting pids test from MonitorTest

When starting a monitor with a PID that doesn't exist, the DOWN message
is immediately sent. This is a core language feature which doesn't have
any implementation on our end, so it doesn't need a test.

* MonitorTest does not match assertions exactly

Previously, tests in MonitorTest checked the exact result of the
monitors in its internal state. This patch makes sure to only test
what's asserted, meaning it only checks if a monitor is added, and if a
monitor is only added once.