Reimplement brittle probes tests (#821)
* Replace probes tests for (un)registering probes

Expose Probes.probes/0 to inspect the current list of probes, and use
that to test if probes can be registered and unregistered.

* Replace test to call the probe in ProbesTest

This new test uses FakeProbe like the previous version, but makes sure
all setip is done in the setup block.

* Replace overridden probes test

Instead of calling the probe multiple times with delays, the new test in
this patch simply makes sure the function is overridden and the state is
updated when a probe is replaced.

* Replace test with disabled probes

Use setup_with_config/0 to disable probes, then repeatedly check to make
sure the probe is not called.

* Don't use FakeProbe in new ProbesTest tests

Instead of using FakeProbe, which has an internal state because it's and
Agent, use a simple probe with a 0 state and a function that increments
the state whenever it's called.

* Replace the failing probe test

We're testing to make sure the probe doesn't take the whole probe runner
down, not wether the function is actually called as that's done
elsewhere.

* Explicitly run the probes in ProbesTest

Instead of waiting for the probes to be run automatically, send
:run_probes to the Probes process to make them run for more stable
tests.

* Remove FakeProbe test module

Since it's replaced by using internal state in ProbesTest, the FakeProbe
module is no longer used, so it can be removed.

* Don't update state in overridden probe test

For the test that checks if a probe's state is overridden, we check if
the state matches the second one passed. We do that by setting the state
to 10, and then setting it to 0.

Previously, the function passed was the same as in the other tests,
meaning it incremented the state for every tick. Since the probes are
running during the tests, there was a small chance the state was updated
after setting the second state and before the assertion.

This patch passes a function that returns an atom, but does not update
the state.