Return state in overridden probe test
Previously, the functions used in the overridden probes test didn't take
an argument, so they returned nil. This patch makes sure both passed
functions take an argument (which is the state) and return it, so the
state doesn't get replaced with a nil.

This fixes the intermittently faling test[1] caused by situations where
the probes were run before the assertion.

[1]: https://appsignal.semaphoreci.com/workflows/50b9faab-6061-4a2b-9b60-172e6eb5ab2c?pipeline_id=585035c8-b553-40c6-8ad9-3aadfe060cd4