Reimplement ErlangProbe tests (#822)
* Add test to make sure the Erlang probe is added

The library automatically adds the Erlang probe, but there was no direct
assertion to check that because the Probes module didn't expose its
current state. This patch adds a check to make sure it's added.

* Move run_probes/0 to Utils

Since the run_probes/0 function is now used by both the Probes and the
ErlangProbes tests, move it to Utils so it doesn't have
to be defined twice.

* Simplify scheduler utilization metrics tests

Instead of asserting on all values, check if the data is set on
subsequent runs, which is what this test should be asserting.

* Unregister Erlang probe in ErlangProbe tests

To make sure the probe is not called through the probes during the
tests, disable the Erlang probe before each test, and reenable it after.

* Move call to set_gauge/3 to ErlangProbe.call/1

To make testing easier, move the call to set_gauge/3 to one place. This
way, the data that's passed to the gauges can be collected before its
set.

* Expose metrics/2 and test its return values

Instead of setting all values through set_gauge/3 and then checking
them, expose the metrics/2 function and test the data that's to be set
directly.

* Don't use FakeAppsignal in ErlangProbeTest

Remove the faked module and only check the values extracted from
Erlang's metrics. This leaves a gap between extracting the data and
sending it to the extension for now.