Fix brittleness in Oban tests (#825)
* Move setup for instrument_oban test to setup block

Instead of detaching the Oban handler in the test, do so in the setup
block and reattach it in an on_exit. That way, the handler is
reattached, even if the test fails.

* Don't start FakeAppsignal if not used in ObanTest

FakeAppsignal's add/2 function checks if the agent is running, and is a
noop if not, so it can be called if we don't care about the result.

* Combine Oban setup blocks into a single one

Instead of having a setup block for every describe, have one main block
to start all faked modules.

* Test the correct span closing in ObanTest

Instead of asserting that one span is closed, check to make sure the
started span was closed. This is more specifically testing what needs to
be tested, and not failing because some other span was also closed in
the same time frame.