Logger backend (#829)
* Expose Appsignal.Logger.log/4

Previously, Appsignal.Logger only exposed the named functions (like
debug/3). Since we need the log/4 function to provide a backend, this
patch eposes it.

* Add Appsignal.Logger.Backend

Appsignal.Logger.Backend is attached to Elixir's logger to send logs to
AppSignal. Use Logger.add_backend/1 to set it up:

    Logger.add_backend({Appsignal.Logger.Backend, "phoenix"})

The second argument in the tuple is the log group name, which could be
"phoenix" for a Phoenix application, for example.

* Add changeset to describe logger backend

> Add Logger backend to redirect Elixir logs to AppSignal.

* Update lib/appsignal/logger/backend.ex

Co-authored-by: Noemi <45180344+unflxw@users.noreply.github.com>

* Use keyword list in Logger.Backend

Instead of taking a string as the log group, take a keyword list with a
single key, for possible extension later.

* Handle :warn warnings in Appsignal.Logger

Deprecated, but still used in older versions of Elixir, warnings named
:warn previously broke the logger. This patch accepts them as warnings.

* Handle unexpected log levels

Although this shouldn't happen, not finding a severity in the list
caused Appsignal.Logger.Backend to crash. This patch adds a rescue
clause to turn everything that doesn't match the list to a severity of
3 (info).

---------

Co-authored-by: Noemi <45180344+unflxw@users.noreply.github.com>